# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Support CRM API
  version: 1.0.0
  description: Support CRM platform API with integrated automation, reporting, and external integrations
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.support.example.com/api/v1
    description: Production
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff session token (via auth/login)
    portalToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Portal account token (via portal/auth/login)
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Integration API token (machine-to-machine via api-tokens)
...
  schemas:
...
    BilingualString:
      type: object
      properties:
        en:
          type: string
        ar:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_FAILED
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - TOO_MANY_REQUESTS
                - INTERNAL_SERVER_ERROR
                - SERVICE_UNAVAILABLE
            message:
              type: string
            request_id:
              type: string
              format: uuid
            field_errors:
              type: [object, 'null']
              additionalProperties:
                type: array
                items:
                  type: string
            meta:
              type: [object, 'null']
...
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /departments/{department}:
    get:
      tags: [Organization]
      operationId: getDepartment
      summary: Get department details
      security:
        - bearer: []
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Department details

    put:
      tags: [Organization]
      operationId: putDepartment
      summary: Update department
      security:
        - bearer: []
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Department updated

    delete:
      tags: [Organization]
      operationId: deleteDepartment
      summary: Delete department
      security:
        - bearer: []
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Department deleted

    patch:
      tags: [Organization]
      operationId: patchDepartment
      summary: Partially update department
      description: All fields optional; UpdateDepartmentRequest applies no `required` rules.
      security:
        - bearer: []
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/BilingualString'
                code:
                  type: string
                  pattern: '^[a-z0-9-]{2,32}$'
      responses:
        '200':
          description: Department updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        '404':
          $ref: '#/components/responses/NotFound'
...
  /teams/{team}:
    get:
      tags: [Organization]
      operationId: getTeam
      summary: Get team details
      security:
        - bearer: []
      parameters:
        - name: team
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team details

    put:
      tags: [Organization]
      operationId: putTeam
      summary: Update team
      security:
        - bearer: []
      parameters:
        - name: team
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Team updated

    delete:
      tags: [Organization]
      operationId: deleteTeam
      summary: Delete team
      security:
        - bearer: []
      parameters:
        - name: team
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Team deleted

    patch:
      tags: [Organization]
      operationId: patchTeam
      summary: Partially update team
      description: All fields optional; UpdateTeamRequest applies no `required` rules.
      security:
        - bearer: []
      parameters:
        - name: team
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/BilingualString'
                code:
                  type: string
                  pattern: '^[a-z0-9-]{2,32}$'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        '404':
          $ref: '#/components/responses/NotFound'
...